Add transactions to deletes.

+++ src/database/MySqlCrud.java
@@ -418,19 +418,48 @@ public class MySqlCrud extends StorageCrud {
      */
     @Override
     public boolean deleteItem(int itemId) {
+        if (!storageService.startTransaction()) {
+            return false; // fail to start transaction
+        }
         // Delete the item from the "Item" table where the ItemId matches the provided
         // itemId.
-        return storageService.delete("Item", "ItemId", itemId);
+        boolean result = storageService.delete("Item", "ItemId", itemId);
+        if (result) {
+            storageService.commitTransaction();
+        } else {
+            storageService.abortTransaction();
+        }
+        return result;
     }
 
     @Override
     public boolean deleteBundle(int bundleId) {
-        return storageService.delete(Bundle.TABLE_NAME, Bundle.BUNDLE_ID_KEY, bundleId);
+        if (!storageService.startTransaction()) {
+            return false; // fail to start transaction
+        }
+
+        boolean result = storageService.delete(Bundle.TABLE_NAME, Bundle.BUNDLE_ID_KEY, bundleId);
+        if (result) {
+            storageService.commitTransaction();
+        } else {
+            storageService.abortTransaction();
+        }
+        return result;
     }
 
     @Override
     public boolean deleteCategory(int categoryId) {
-        return storageService.delete(Category.TABLE_NAME, Category.CATEGORY_ID_KEY, categoryId);
+        if (!storageService.startTransaction()) {
+            return false; // fail to start transaction
+        }
+
+        boolean result = storageService.delete(Category.TABLE_NAME, Category.CATEGORY_ID_KEY, categoryId);
+        if (result) {
+            storageService.commitTransaction();
+        } else {
+            storageService.abortTransaction();
+        }
+        return result;
     }
 
 }
